move dllload and Filherper to shared lib
