+++ Samples/SampleSharedLib/SampleSharedLib/FileHelper.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <string>
+#include <Windows.h>
+namespace FileHelper
+{
+  std::wstring GetModulePath();
+}
