+++ Samples/SampleSharedLib/SampleSharedLib/pch.cpp
@@ -1,0 +1,1 @@
+#include "pch.h"
