+++ Samples/SampleSharedLib/SampleSharedLib/pch.h
@@ -1,0 +1,7 @@
+#ifndef PCH_H
+#define PCH_H
+
+#include <string>
+#include <fstream>
+
+#endif //PCH_H
